bind: ruby: working but benchmark is breaking
The marking and interior mutability don't seem to work well together.

If the GC kicks the mark phase, and we iterate over the tree to mark all
the descendants, we might, just might, hit the case where we're:

  1. Borrowing mutably to insert.
  2. Borrowing immutably to iterate and mark.

So in my benchmarks, the magic number is ~300 inserts of (int, int).